# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
  dto.AreaCreateDTO:
    properties:
      name:
        type: string
    required:
    - name
    type: object
  dto.AreaResponseDTO:
    properties:
      id:
        type: integer
      name:
        type: string
      points:
        items:
          $ref: '#/definitions/dto.PointResponseDTO'
        type: array
    type: object
  dto.AreaUpdateDTO:
    properties:
      name:
        type: string
    type: object
  dto.ErrorResponse:
    properties:
      error:
        type: string
      message:
        type: string
    type: object
  dto.LoginDTO:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  dto.LoginResponse:
    properties:
      message:
        type: string
      token:
        type: string
    type: object
  dto.PhotoDTO:
    properties:
      url:
        type: string
    type: object
  dto.PointDTO:
    properties:
      latitude:
        type: string
      longitude:
        type: string
    required:
    - latitude
    - longitude
    type: object
  dto.PointResponseDTO:
    properties:
      area_id:
        type: integer
      created_at:
        type: string
      id:
        type: integer
      latitude:
        type: string
      longitude:
        type: string
      updated_at:
        type: string
    type: object
  dto.PointUpdateDTO:
    properties:
      area_id:
        type: integer
      latitude:
        type: string
      longitude:
        type: string
    type: object
  dto.ReportCreatedDTO:
    properties:
      description:
        type: string
      latitude:
        type: number
      longitude:
        type: number
      name:
        type: string
      photos:
        items:
          type: string
        type: array
      report_status_id:
        type: integer
      report_type_id:
        type: integer
      user_id:
        type: integer
    type: object
  dto.ReportResponseDTO:
    properties:
      message:
        type: string
      report:
        $ref: '#/definitions/model.Report'
    type: object
  dto.ReportUpdateDTO:
    properties:
      description:
        type: string
      latitude:
        type: number
      longitude:
        type: number
      name:
        type: string
      photos:
        items:
          $ref: '#/definitions/dto.PhotoDTO'
        type: array
      report_id:
        type: integer
      report_status_id:
        type: integer
      report_type_id:
        type: integer
      user_id:
        type: integer
    type: object
  dto.SuccessResponse:
    properties:
      message:
        type: string
    type: object
  dto.UserCreatedDTO:
    properties:
      email:
        type: string
      firstname:
        type: string
      lastname:
        type: string
      password:
        type: string
      role:
        type: integer
      username:
        type: string
    required:
    - email
    - firstname
    - lastname
    - password
    - username
    type: object
  dto.UserResponseDTO:
    properties:
      createdAt:
        type: string
      email:
        type: string
      firstname:
        type: string
      id:
        type: integer
      lastname:
        type: string
      role:
        type: string
      username:
        type: string
    required:
    - createdAt
    - email
    - firstname
    - id
    - lastname
    - role
    - username
    type: object
  dto.UserUpdateDTO:
    properties:
      email:
        type: string
      firstname:
        type: string
      lastname:
        type: string
      userid:
        type: string
      username:
        type: string
    required:
    - userid
    type: object
  model.Comment:
    properties:
      content:
        type: string
      created_at:
        type: string
      id:
        type: integer
      report_id:
        type: integer
      user:
        $ref: '#/definitions/model.User'
      user_id:
        type: integer
    type: object
  model.Photo:
    properties:
      created_at:
        type: string
      id:
        type: integer
      image:
        type: string
      report:
        $ref: '#/definitions/model.Report'
      reportID:
        type: integer
    type: object
  model.Report:
    properties:
      comments:
        items:
          $ref: '#/definitions/model.Comment'
        type: array
      description:
        type: string
      id:
        type: integer
      last_update:
        type: string
      latitude:
        type: number
      longitude:
        type: number
      name:
        type: string
      photos:
        items:
          $ref: '#/definitions/model.Photo'
        type: array
      report_date:
        type: string
      report_status:
        $ref: '#/definitions/model.ReportStatus'
      report_status_id:
        type: integer
      report_type:
        $ref: '#/definitions/model.ReportType'
      report_type_id:
        type: integer
      updates:
        items:
          $ref: '#/definitions/model.ReportUpdate'
        type: array
      user:
        $ref: '#/definitions/model.User'
      user_id:
        type: integer
    type: object
  model.ReportStatus:
    properties:
      id:
        type: integer
      status:
        type: string
    type: object
  model.ReportType:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
  model.ReportUpdate:
    properties:
      id:
        type: integer
      report:
        $ref: '#/definitions/model.Report'
      report_id:
        type: integer
      report_status:
        $ref: '#/definitions/model.ReportStatus'
      report_status_id:
        type: integer
      updated_at:
        type: string
    type: object
  model.Role:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
  model.User:
    properties:
      creationDate:
        type: string
      email:
        type: string
      first_name:
        type: string
      id:
        type: integer
      last_name:
        type: string
      password:
        type: string
      role:
        $ref: '#/definitions/model.Role'
      role_id:
        type: integer
      username:
        type: string
    type: object
externalDocs:
  description: OpenAPI
  url: https://swagger.io/resources/open-api/
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server celler server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Swagger Example API
  version: "1.0"
paths:
  /area:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.AreaResponseDTO'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Get all areas
    post:
      consumes:
      - application/json
      parameters:
      - description: Area data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.AreaCreateDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.AreaResponseDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Create a new area
  /area/{area_id}/point:
    post:
      consumes:
      - application/json
      description: Creates a new point associated with an area.
      parameters:
      - description: Area ID
        in: path
        name: area_id
        required: true
        type: integer
      - description: Point data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.PointDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.PointResponseDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Create Point
      tags:
      - Point
  /area/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Area ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Delete an area
    get:
      consumes:
      - application/json
      parameters:
      - description: Area ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.AreaResponseDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Get an area by ID
    put:
      consumes:
      - application/json
      parameters:
      - description: Area ID
        in: path
        name: id
        required: true
        type: integer
      - description: Area data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.AreaUpdateDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.AreaResponseDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Update an area
  /area/{id}/point:
    get:
      consumes:
      - application/json
      description: Retrieves all points associated with an area by its ID.
      parameters:
      - description: Area ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.PointResponseDTO'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Get Points by Area ID
      tags:
      - Point
  /dashboard:
    get:
      consumes:
      - application/json
      description: Returns the user dashboard
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.User'
      security:
      - ApiKeyAuth: []
      summary: User dashboard
      tags:
      - users
  /point/{id}:
    delete:
      consumes:
      - application/json
      description: Deletes a point associated with an area by its ID.
      parameters:
      - description: Point ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Delete Point
      tags:
      - Point
    put:
      consumes:
      - application/json
      description: Updates a point associated with an area by its ID.
      parameters:
      - description: Point ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated point data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.PointUpdateDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.PointResponseDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Update Point
      tags:
      - Point
  /report:
    post:
      consumes:
      - application/json
      description: creates report in database
      parameters:
      - description: query params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.ReportCreatedDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ReportResponseDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ReportResponseDTO'
      summary: Create Report
      tags:
      - Report
  /report/:
    put:
      consumes:
      - application/json
      description: updates report in database
      parameters:
      - description: Report Update Params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.ReportUpdateDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ReportResponseDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ReportResponseDTO'
      summary: Update Report
      tags:
      - Report
  /report/{id}:
    get:
      consumes:
      - application/json
      description: gets report from database
      parameters:
      - description: report ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ReportResponseDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ReportResponseDTO'
      summary: Get Report
      tags:
      - Report
  /report/user/{id}:
    get:
      consumes:
      - application/json
      description: Get all reports from a user by user ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Report'
            type: array
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ReportResponseDTO'
      summary: Get Reports by User ID
      tags:
      - Report
  /user:
    get:
      consumes:
      - application/json
      description: returns all users in db
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.UserResponseDTO'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.UserResponseDTO'
      summary: Get All User
      tags:
      - User
  /user/{id}:
    get:
      consumes:
      - application/json
      description: Returns a user from the database by their ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.UserResponseDTO'
        "404":
          description: Not Found
          schema:
            type: string
      summary: Get User by ID
      tags:
      - User
  /user/edit:
    put:
      consumes:
      - application/json
      description: edits user information in database
      parameters:
      - description: query params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UserUpdateDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.UserResponseDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.UserResponseDTO'
      summary: Edit User
      tags:
      - User
  /user/login:
    post:
      consumes:
      - application/json
      description: returns token for login
      parameters:
      - description: query params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.LoginDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.LoginResponse'
      summary: Login
      tags:
      - User
  /user/register:
    post:
      consumes:
      - application/json
      description: registers user in database
      parameters:
      - description: query params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UserCreatedDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.UserResponseDTO'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.UserResponseDTO'
      summary: Register
      tags:
      - User
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"